Refactor HintRALStatement and HintRALStatementExecutor
