+++ shardingsphere-features/shardingsphere-readwrite-splitting/shardingsphere-readwrite-splitting-distsql/shardingsphere-readwrite-splitting-distsql-statement/src/main/java/org/apache/shardingsphere/readwritesplitting/distsql/parser/statement/hint/SetReadwriteSplittingHintStatement.java
@@ -19,14 +19,14 @@ package org.apache.shardingsphere.readwritesplitting.distsql.parser.statement.hi
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.distsql.parser.statement.ral.common.HintDistSQLStatement;
+import org.apache.shardingsphere.distsql.parser.statement.ral.HintRALStatement;
 
 /**
  * Set readwrite-splitting hint statement.
  */
 @RequiredArgsConstructor
 @Getter
-public final class SetReadwriteSplittingHintStatement extends HintDistSQLStatement {
+public final class SetReadwriteSplittingHintStatement extends HintRALStatement {
     
     private final String source;
 }
