+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-statement/src/main/java/org/apache/shardingsphere/sharding/distsql/parser/statement/hint/SetShardingHintDatabaseValueStatement.java
@@ -19,14 +19,14 @@ package org.apache.shardingsphere.sharding.distsql.parser.statement.hint;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.distsql.parser.statement.ral.common.HintDistSQLStatement;
+import org.apache.shardingsphere.distsql.parser.statement.ral.HintRALStatement;
 
 /**
  * Set sharding hint database value statement.
  */
 @RequiredArgsConstructor
 @Getter
-public final class SetShardingHintDatabaseValueStatement extends HintDistSQLStatement {
+public final class SetShardingHintDatabaseValueStatement extends HintRALStatement {
     
     private final String shardingValue;
 }
